refactor(web): set confirm door button on every page
